Actualización de los controllers y repositories

+++ safetrip-backend/src/main/java/com/safetripbackend/entity/Cities.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Entity
 @AllArgsConstructor
@@ -12,8 +14,7 @@ import lombok.NoArgsConstructor;
 public class Cities {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     private String name;
     private String location;
     private String sights;
