+++ safetrip-backend/src/main/java/com/safetripbackend/entity/Users.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Entity
 @AllArgsConstructor
@@ -12,9 +14,8 @@ import lombok.NoArgsConstructor;
 public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-    private String email;
-    private String password;
+    private int id;
+    private String name; // Ejemplo de atributo de tipo cadena
+    private String email; // Ejemplo de atributo de tipo cadena
+    private String password; // Ejemplo de atributo de tipo cadena
 }
